// bkupcli/src/app/model/pictureandqrinfo.model.ts
export class PictureAndQRInfo {

    pictureAndQrSeqId: string;
    pictureAndQrTnxSeqId: string;
    entity: string;
    productCode: string;
    prodRefId: string;
    businessDate: Date;
    trackingCode: string;
    trackingFileName: string;
    trackingFileLocation: string;
    trackingCreateDate: Date;
    verifiedDate: Date;
    verifiedCode: string;
    verifiedFileName: string;
    verifiedFileLocation: string;
    pictureCode: string;
    pictureFileName: string;
    pictureFileLocation: string;
    pictureCaptureDate: Date;
    imageSrc: string;
    qrSrc: string;

}